api/requirements.txt
annotated_types==0.6.0
Django==5.0.3
django_environ==0.11.2
pydantic==2.6.4
pydantic_core==2.16.3
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg==3.1.18
psycopg-binary==3.1.18